+++ Api-Agendate-App/DTOs/Usuarios/UsuarioDTO.cs
@@ -6,6 +6,7 @@ namespace Api_Agendate_App.DTOs
 {
     public class UsuarioDTO
     {
+        [JsonIgnore]
         [Required(ErrorMessage = "El campo No puede ser nulo o vacio")]
         public int Id { get; set; }
